requirements.txt
keras
numpy
urllib3
matplotlib
seaborn
scikit-learn
opencv-python
gensim
pillow
regex
tensorflow-gpu==1.6
